+++ CasualConversation/Targets/Domain/Sources/Service/AudioStatus.swift
@@ -1,0 +1,16 @@
+//
+//  AudioStatus.swift
+//  Domain
+//
+//  Created by Yongwoo Marco on 2022/08/17.
+//  Copyright © 2022 pseapplications. All rights reserved.
+//
+
+import Foundation
+
+public enum AudioStatus {
+	case stopped
+	case paused
+	case recording
+	case playing
+}
